// gui_src/vendor/docx-editor/core/types/content/trackedChange.ts
export interface RevisionInfo {
  revisionId: number;
  author: string;
  date: string | null;
}